Register check requirements
Requirements can be any of a subset of possible targets:
- `Table(name)`
- `Column(name)`
- `Column(name, table)`

These are specified either explicitly, or computed from a function
evaluted on the keyword argument values of the check.

A missing requirement now results in the check being skipped rather
than an error (presumably) being raised within the check function.