Split result with unequal test and input indices
This scenario can occur in presence tests:
something is absent in the input and reported as invalid.

NOTE: It may be more robust to require that test index is a subset
of the input index, and use custom messages to communicate the names
of the missing elements. For example:

`has_column(['x', 'y']) -> "Missing columns ['y']"`